py-requirements.txt
spotipy
pandas
python-dotenv
ipykernel
numpy
matplotlib
networkx
